build: Set python_requires to Python 2.7 or 3.5+
Having a lower bound on python_requires is critical to ensure that pip
can properly resolve and install the package in the future. If there is no
python_requires then when compatibility with a version is broken (e.g. the
inevitable and hopefully near-term dropping of Python 2) pip will not know
that the version is broken on that Python and will attempt to install a broken
release. Setting a lower bond on python_requires and carefully monitoring
and bumping it through testing is critical.